requirements.txt
# Core ML and Deep Learning
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Configuration Management
PyYAML>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Experiment Tracking
mlflow>=2.5.0
wandb>=0.15.0
tensorboard>=2.13.0

# Model Explainability
lime>=0.2.0
shap>=0.42.0
grad-cam>=1.4.0

# Image Processing and Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Data Validation
pydantic>=2.0.0
cerberus>=1.3.0

# File Handling
pathlib2>=2.3.0